CommentsButtonsList at TransactionAddForm fixed
